Rename placecell module to camap
