+++ .github/workflows/test.yml
@@ -37,7 +37,7 @@ jobs:
         run: uv pip install --no-binary oasis-deconv oasis-deconv
 
       - name: Run tests with coverage
-        run: uv run pytest tests/ -v --cov=placecell --cov-report=xml --cov-report=term
+        run: uv run pytest tests/ -v --cov=camap --cov-report=xml --cov-report=term
 
       - name: Run notebook smoke tests
         run: uv run pytest --nbmake notebook/view_results_arena.ipynb notebook/view_results_maze.ipynb -v
