+++ camap/analysis/__init__.py
@@ -1,6 +1,6 @@
 """Spatial analysis functions for place cells."""
 
-from placecell.analysis.pvo_1d import (
+from camap.analysis.pvo_1d import (
     compute_dataset_arm_pvo,
     plot_arm_pvo_grid,
 )
