+++ camap/config/example_arena_config.yaml
