+++ camap/config/example_maze_config.yaml
