+++ camap/config/example_zone_graph.yaml
