+++ camap/dataset/__init__.py
@@ -1,13 +1,13 @@
 """Dataset classes for place cell analysis."""
 
-from placecell.dataset.arena import ArenaDataset
-from placecell.dataset.base import (
+from camap.dataset.arena import ArenaDataset
+from camap.dataset.base import (
     BasePlaceCellDataset,
     StabilitySplitResult,
     UnitResult,
     unique_bundle_path,
 )
-from placecell.dataset.maze import MazeDataset
+from camap.dataset.maze import MazeDataset
 
 __all__ = [
     "ArenaDataset",
