+++ camap/geometry.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import numpy as np
 import yaml
 
-from placecell.log import init_logger
+from camap.log import init_logger
 
 logger = init_logger(__name__)
