+++ docs/installation.md
@@ -3,14 +3,14 @@
 > **Note:** PyPI distribution is planned. For now, install directly from GitHub.
 
 ```bash
-pip install git+https://github.com/miniscope/placecell.git
+pip install git+https://github.com/miniscope/camap.git
 ```
 
 ## Install oasis-deconv (required for deconvolution)
 
-placecell uses [`oasis-deconv`](https://github.com/j-friedrich/OASIS) for
+camap uses [`oasis-deconv`](https://github.com/j-friedrich/OASIS) for
 its deconvolution step. It is **required** for the full pipeline but
-**not bundled** with placecell — PyPI wheel coverage is patchy (arm64
+**not bundled** with camap — PyPI wheel coverage is patchy (arm64
 macOS only), so we leave the install path to you:
 
 ```bash
@@ -27,7 +27,7 @@ If the source build fails, that is an upstream issue — see the
 ## Development
 
 ```bash
-git clone https://github.com/miniscope/placecell.git
-cd placecell
+git clone https://github.com/miniscope/camap.git
+cd camap
 uv sync --extra all
 ```
