+++ docs/modules.md
@@ -5,16 +5,16 @@
    :toctree: _generated
    :recursive:
 
-   placecell.dataset
-   placecell.dataset.base
-   placecell.dataset.arena
-   placecell.dataset.maze
-   placecell.analysis.spatial_2d
-   placecell.analysis.spatial_1d
-   placecell.behavior
-   placecell.config
-   placecell.loaders
-   placecell.maze_helper
-   placecell.neural
-   placecell.visualization
+   camap.dataset
+   camap.dataset.base
+   camap.dataset.arena
+   camap.dataset.maze
+   camap.analysis.spatial_2d
+   camap.analysis.spatial_1d
+   camap.behavior
+   camap.config
+   camap.loaders
+   camap.maze_helper
+   camap.neural
+   camap.visualization
 ```
