+++ docs/pipeline.md
@@ -13,7 +13,7 @@ Both `ArenaDataset` (2D) and `MazeDataset` (1D) implement the same abstract pipe
 5. `match_events()` — interpolate behavior onto neural timestamps, compute speed, build canonical table
 6. `compute_occupancy()` — spatial occupancy map (2D histogram or 1D bins)
 7. `analyze_units()` — per-unit rate map, spatial information, stability, place fields
-8. `save_bundle()` — write `.pcellbundle` directory with config, arrays, parquets, figures, and a `metadata.json` that records both the bundle schema version and the `placecell` package version (via `hatch-vcs`, so it includes the git SHA of the build).
+8. `save_bundle()` — write `.pcellbundle` directory with config, arrays, parquets, figures, and a `metadata.json` that records both the bundle schema version and the `camap` package version (via `hatch-vcs`, so it includes the git SHA of the build).
 
 Steps 3–4 are gated on which blocks the data config carries: `preprocess_behavior()` is a no-op when no `behavior:` block is present, `deconvolve()` is a no-op when no `neural:` block is present, and the place-cell steps 5–7 require both. This makes neural-only (load → deconvolve → save) and behavior-only (load → preprocess → save) workflows valid out of the box.
 
@@ -67,7 +67,7 @@ Saves a copy of the raw trajectory in `trajectory_raw`, then applies geometric c
 
 #### `MazeDataset` (1D)
 
-`MazeDataset.load()` reads a `zone_tracking` CSV that maps each **neural** frame to a zone label and an arm-relative position. If the CSV is missing, `load()` runs [Zone Detection](#zone-detection-maze-only) automatically; pass `--force-redetect` to `placecell analysis` (or `force_redetect=True` to `MazeDataset.load()`) to refresh it.
+`MazeDataset.load()` reads a `zone_tracking` CSV that maps each **neural** frame to a zone label and an arm-relative position. If the CSV is missing, `load()` runs [Zone Detection](#zone-detection-maze-only) automatically; pass `--force-redetect` to `camap analysis` (or `force_redetect=True` to `MazeDataset.load()`) to refresh it.
 
 `preprocess_behavior()` then operates on the already-neural-rate trajectory and:
 
@@ -86,11 +86,11 @@ Requires both a `neural:` and a `behavior:` block — raises with a targeted mes
 4. Drop neural frames with no behavior coverage (outside behavior time window).
 5. Derive the speed-filtered view (`trajectory_filtered` / `trajectory_1d_filtered`) and the long-format `event_place` / `event_index` tables.
 
-The hard-error guardrail in :func:`placecell.behavior.interpolate_behavior_onto_neural` refuses to upsample when `neural_fps > 5 * behavior_fps`, since per-frame jitter would dominate.
+The hard-error guardrail in :func:`camap.behavior.interpolate_behavior_onto_neural` refuses to upsample when `neural_fps > 5 * behavior_fps`, since per-frame jitter would dominate.
 
 ### Zone Detection (maze only)
 
-Zone detection projects raw DLC `(x, y)` onto the maze graph **at the neural sample rate**. It runs automatically from `MazeDataset.load()` when the `zone_tracking` CSV is missing, and can also be invoked directly via `placecell detect-zones -d data_paths.yaml` (which additionally exports a validation video).
+Zone detection projects raw DLC `(x, y)` onto the maze graph **at the neural sample rate**. It runs automatically from `MazeDataset.load()` when the `zone_tracking` CSV is missing, and can also be invoked directly via `camap detect-zones -d data_paths.yaml` (which additionally exports a validation video).
 
 1. Load raw DLC `(x, y)` from `behavior_position.csv`.
 2. **Hampel jump removal** on the raw trajectory at behavior rate, using `zone_detection.hampel_window_frames` and `zone_detection.hampel_n_sigmas`.
@@ -278,7 +278,7 @@ behavior:
 ```
 :::
 
-`placecell` reads the DLC scorer name from the CSV header and does not require a fixed scorer string such as `3DMazeTrack` or `FuzzyTrack`.
+`camap` reads the DLC scorer name from the CSV header and does not require a fixed scorer string such as `3DMazeTrack` or `FuzzyTrack`.
 
 ### Analysis Config
