+++ docs/quickstart.md
@@ -1,6 +1,6 @@
 # Quickstart
 
-`placecell` supports two workflows:
+`camap` supports two workflows:
 
 - `arena` for 2D open-field analysis
 - `maze` for 1D arm/graph-based analysis
@@ -73,11 +73,11 @@ behavior:
 ```
 :::
 
-`placecell` is scorer-agnostic for DLC-style CSVs; configure the correct `bodypart`, and the scorer name is read from the file header.
+`camap` is scorer-agnostic for DLC-style CSVs; configure the correct `bodypart`, and the scorer name is read from the file header.
 
 ### 1b. Maze: zone detection
 
-For maze sessions, the analysis pipeline projects the trajectory onto a zone graph at the neural sample rate. `placecell analysis` will run zone detection automatically on first use, so for a basic run you can skip ahead. See [CLI Workflows](workflows.md) for the `define-zones` → `detect-zones` flow when you want to author the zone graph or iterate on detection parameters.
+For maze sessions, the analysis pipeline projects the trajectory onto a zone graph at the neural sample rate. `camap analysis` will run zone detection automatically on first use, so for a basic run you can skip ahead. See [CLI Workflows](workflows.md) for the `define-zones` → `detect-zones` flow when you want to author the zone graph or iterate on detection parameters.
 
 ### 2. Create analysis config
 
@@ -132,13 +132,13 @@ behavior:
 ### 3. Run the analysis
 
 ```bash
-placecell analysis -c config.yaml -d data_paths.yaml
+camap analysis -c config.yaml -d data_paths.yaml
 ```
 
 Or via Python:
 
 ```python
-from placecell.dataset import BasePlaceCellDataset
+from camap.dataset import BasePlaceCellDataset
 
 ds = BasePlaceCellDataset.from_yaml("config.yaml", "data_paths.yaml")
 ds.load()
