+++ tests/test_regression_1d.py
@@ -5,7 +5,7 @@ compares every output against a saved reference bundle.
 
 To regenerate the reference bundle with low ``n_shuffles`` in the config::
 
-    placecell analysis -c config.yaml -d data.yaml -o tests/assets/regression_1d/reference -y --subset-units 10 --subset-frames 10000
+    camap analysis -c config.yaml -d data.yaml -o tests/assets/regression_1d/reference -y --subset-units 10 --subset-frames 10000
 """
 
 import shutil
@@ -15,8 +15,8 @@ from pathlib import Path
 import numpy as np
 import pytest
 
-from placecell.dataset.base import BasePlaceCellDataset
-from placecell.dataset.maze import MazeDataset
+from camap.dataset.base import BasePlaceCellDataset
+from camap.dataset.maze import MazeDataset
 
 REGRESSION_DIR = Path(__file__).parent / "assets" / "regression_1d"
